Refactoring for release
