+++ bundles/binding/org.openhab.binding.ebus/.project
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
 	<name>org.openhab.binding.ebus</name>
-	<comment>This is the NTP binding of the open Home Automation Bus (openHAB)</comment>
+	<comment>This is the eBus binding of the open Home Automation Bus (openHAB)</comment>
 	<projects>
 	</projects>
 	<buildSpec>
