+++ bundles/binding/org.openhab.binding.ebus/OSGI-INF/ebusbinding.xml
@@ -10,7 +10,7 @@
 
 -->
 <scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" activate="activate" deactivate="deactivate" name="org.openhab.binding.ebus">
-	<implementation class="org.openhab.binding.ebus.EBusBinding" />
+	<implementation class="org.openhab.binding.ebus.internal.EBusBinding" />
 
 	<service>
 		<provide interface="org.osgi.service.cm.ManagedService" />
