--- bundles/binding/org.openhab.binding.ebus/src/main/java/org/openhab/binding/ebus/Activator.java
@@ -1,45 +1,0 @@
-/**
-* Copyright (c) 2010-2014, openHAB.org and others.
-*
-* All rights reserved. This program and the accompanying materials
-* are made available under the terms of the Eclipse Public License v1.0
-* which accompanies this distribution, and is available at
-* http://www.eclipse.org/legal/epl-v10.html
-*/
-package org.openhab.binding.ebus;
-
-import org.osgi.framework.BundleActivator;
-import org.osgi.framework.BundleContext;
-
-/**
-* @author Christian Sowada
-* @since 1.7.0
-*/
-public class Activator implements BundleActivator {
-
-	private static BundleContext context;
-
-	/**
-	 * Returns the bundle context
-	 * @return bundle context
-	 */
-	static BundleContext getContext() {
-		return context;
-	}
-
-	/*
-	 * (non-Javadoc)
-	 * @see org.osgi.framework.BundleActivator#start(org.osgi.framework.BundleContext)
-	 */
-	public void start(BundleContext bundleContext) throws Exception {
-		Activator.context = bundleContext;
-	}
-
-	/*
-	 * (non-Javadoc)
-	 * @see org.osgi.framework.BundleActivator#stop(org.osgi.framework.BundleContext)
-	 */
-	public void stop(BundleContext bundleContext) throws Exception {
-		Activator.context = null;
-	}
-}
