+++ bundles/binding/org.openhab.binding.ebus/src/main/resources/common-configuration.json
