+++ bundles/binding/org.openhab.binding.ebus/src/main/resources/testing-configuration.json
