+++ bundles/binding/org.openhab.binding.ebus/src/main/resources/vaillant-configuration.json
