+++ bundles/binding/org.openhab.binding.ebus/src/main/resources/wolf-configuration.json
